Run openstack-ceilometer-central with Pacemaker/Systemd
The resource agent is not working properly on RHEL plateforms because of
systemd.
For now, let's use systemd resource in Pacemaker.
Later, we will refactore all Pacemaker usage to have something
Puppetish.

The idea is for now to close the bug and drop a blocker to deploye this
service.

Closes-bug: #1403131

Change-Id: Ia694c2c460db882ba2f82aedad55c1a77691397f
(cherry picked from commit 1c7669921a4598abf2cf7b6098418c6599f67e59)